Allow makeaclitem() to accept multiple privileges.
Currently makeaclitem() allows only a single privilege at a time,
which works however this is does not allow generating complex ACLs
for a given grantor/grantee entry.

This feature is alreasy present for all the has_foo_privilege() set
of functions. This commit only reuses that infrastructure for
makeaclitem().

Such an enhancement is helpful in situations where aclitem[] needs
to be reconstructed after surgically removing some privileges
related to a grantor or a grantee. Such a scenario came up in a
recent upgrade failure where the database had multiple ACLs
(aclitem[]::TEXT was more than 2000 characters), wherein some of
the grantees were missing in pg_roles. Now although aclexplode()
could be used to remove some grantor / grantees, however it was
non-trivial to reconstruct the same without makeaclitem()
supporting multiple privileges.