Clip first byte of of only single-byte strings.
With this commit text_char() returns the first character of only
the single-byte strings, and returns an error for non-empty multi-byte
strings (alike its i4tochar() counterpart).

Earlier, silently clipping the first byte of multi-byte strings could
end up returning invalid data, which could cause assert(). To make things
worse, reusing this malformed "char" in another string function (such as
right()) could cause an engine crash. This has been so forever, but given
that it is trivial to cause an engine crash, it should be backpatched to
all major versions.